+++ consts/consts.go
@@ -5,10 +5,10 @@ package consts
 
 const (
 	// Version 库版本号，对应 aiotieba.__version__。
-	Version = "1.4.1"
+	Version = "1.5.0"
 
 	// LatestVersion 通常用于大部分API。
-	LatestVersion = "22.6.5.1"
+	LatestVersion = "22.10.1.0"
 	// LegacyVersion 通常用于部分依赖旧版格式的API (`get_threads`...)。
 	LegacyVersion = "12.64.1.1"
 	// ChatVersion 用于聊天 / BLCP 协议。